requirements.txt
# gRPC 基础
grpcio==1.60.0
grpcio-tools==1.60.0

# AI 核心
torch
transformers==4.49.0
pillow
requests

# OCR (飞桨)
paddlepaddle
paddleocr
opencv-python-headless

# Qwen3-VL 依赖
qwen-vl-utils
accelerate>=0.26.0
torchvision

numpy==1.26.4
addict==2.4.0
einops
easydict

mlx-vlm